Saved Posts is successfully moved to seperate component
