+++ src/components/Post/PostList.jsx
@@ -4,53 +4,12 @@ import moment from 'moment';
 import reddit_logo from '../../images/reddit-logo.png'
 import { FaRegBookmark, FaRegCommentAlt, FaRegFlag, FaShare } from 'react-icons/fa'
 import { TbArrowBigTop, TbArrowBigDown } from 'react-icons/tb'
-import axios from 'axios';
 import '../../css/post.css';
-import { toast } from 'react-toastify';
+import SavePosts from './SavePosts';
 
 const PostList = ({ account, posts, community }) => {
 
-const [isSaved, setIsSaved] = useState(false);
-const current_account = JSON.parse(localStorage.getItem('account'))
-
-
-useEffect(() => {
-  const fetchSavedPosts = async () => {
-    // make an API call to fetch the saved posts
-    const response = await fetch(`http://localhost:3000/api/v1/accounts/${account_id}/saved_posts`);
-    const data = await response.json();
-    setIsSaved(data);
-  };
-
-  fetchSavedPosts();
-}, []);
-
-const handleSave = async (postId) => {
-  try {
-    if (isSaved) {
-      await axios.delete(`http://localhost:3000/api/v1/accounts/${current_account.id}/save_posts/${postId}`);
-      console.log("Delete");
-    } else {
-      const response = await fetch(`http://localhost:3000/api/v1/accounts/${current_account.id}/save_posts/`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        
-        body: JSON.stringify({post_id: postId, account_id: current_account.id })
-      });
-      const data =await response.json();
-      if (response.status === 201) {
-        toast.success("Post Saved successfully!");
-      }
-    }
-    setIsSaved(!isSaved);
-  } catch (error) {
-    console.error(error);
-  }
-};
-
-  if (!Array.isArray(posts)) {
+if (!Array.isArray(posts)) {
     return null;
   }
   return (
@@ -108,15 +67,7 @@ const handleSave = async (postId) => {
                     <Link to='' className="list-post-tab">
                       <FaShare /> Share
                     </Link>
-                    <Link to='' className={`list-post-tab ${isSaved ? 'saved' : ''}`} 
-                      onClick={() => handleSave(post.id)}>
-                        {isSaved ? (
-                          <FaRegBookmark style={{ fill: 'grey' }} />
-                         ) : (
-                           <FaRegBookmark />
-                        )}
-                        Save
-                    </Link>
+                    <SavePosts post={post}/>
                     <Link to={`/r/${post.community_id}/p/${post.id}`} className="list-post-tab">
                       <FaRegFlag /> Report
                     </Link>
